+++ NutritionService/Features/MealsRecommendations/GetMealRecommendationsQuery.cs
@@ -3,5 +3,5 @@ using NutritionService.Shared.Response;
 
 namespace NutritionService.Features.MealsRecommendations
 {
-    public record GetMealRecommendationsQuery(string MealType , int pageNumber , int pageSize , int? maxCalories , int? minProtein) : IRequest<ResponseResult<List<MealRecommendationModelView>>>;   
+    public record GetMealRecommendationsQuery(string MealType , int pageNumber , int pageSize , int? maxCalories , double? minProtein) : IRequest<ResponseResult<List<MealRecommendationModelView>>>;   
 }
